pyproject.toml
[project]
name = "stock-analyzer"
version = "0.1.0"
description = "한국 주식 분석 리포트 생성 CLI 애플리케이션"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # setuptools for pkg_resources (required by pykrx)
    "setuptools>=70.0.0",
    # 주가 데이터
    "pykrx>=1.0.45",
    # DART 공시
    "opendartreader>=0.2.1",
    # 기술적 지표
    "pandas-ta>=0.3.14b0",
    # 데이터 처리
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    # 웹 스크래핑
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    # PDF 생성
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    # 차트
    "matplotlib>=3.7.0",
    # 설정
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    # AI
    "openai>=1.0.0",
    # 카카오톡 & Google Drive
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",
]

[project.scripts]
stock-report = "stock_analyzer.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_analyzer"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
